+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### ✨ Added
 
 - `std::string` support in JSON serializer.
+- Native `enum class` support in JSON serializer.
 
 ### 🙌 Improvements